pharmaceutical_quiz/requirements.txt
# Core dependencies for Pharmaceutical Quiz Module
streamlit>=1.28.0
openai>=1.0.0
anthropic>=0.7.0
requests>=2.31.0

# Document processing (updated stack)
PyMuPDF>=1.23.0
python-docx>=0.8.11
python-pptx>=0.6.21
pandas>=2.0.0
openpyxl>=3.1.0

# Advanced text processing
spacy>=3.7.0

# Data visualization
plotly>=5.17.0

# UI enhancements
streamlit-option-menu>=0.3.6
streamlit-lottie>=0.0.5

# Report generation
fpdf2>=2.7.6

# Additional utilities
python-dotenv>=1.0.0
pillow>=10.0.0

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
